init delete users
Demos inicio ao desev do metodo de deleção dos usuários

+++ src/main/java/br/com/sharetips/services/UserService.java
@@ -3,8 +3,10 @@ package br.com.sharetips.services;
 import java.util.List;
 import java.util.Optional;
 
-import br.com.sharetips.entities.dto.LoginUserDTO;
+import br.com.sharetips.entities.Tip;
+import br.com.sharetips.entities.dto.user.UserLoginRequestDTO;
 import br.com.sharetips.exceptions.ResourceNotFoundException;
+import br.com.sharetips.repositories.TipRepository;
 import br.com.sharetips.services.exceptions.DatabaseException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +18,9 @@ import br.com.sharetips.repositories.UserRepository;
 public class UserService {
 	@Autowired
 	private UserRepository repository;
+
+	@Autowired
+	private TipRepository tipRepository;
 	
 	public List<User> findAll() {
 		return repository.findAll();
@@ -37,15 +42,19 @@ public class UserService {
 		}
 	}
 
-	public User login(LoginUserDTO dto) {
+	public User login(UserLoginRequestDTO dto) {
 		Optional<User> user = repository.findByEmailAndPassword(dto.getEmail(), dto.getPassword());
 
 		return user.orElseThrow(() -> new ResourceNotFoundException("Email ou senha incorretos"));
 	}
 
 	public void deleteById(Long id) {
-		findById(id);
-		repository.deleteById(id);
+		User user = findById(id);
+		List<Tip> tips = tipRepository.findByAuthor(user);
+
+		if(tips.isEmpty()) {
+			repository.deleteById(id);
+		}
 	}
 
 	public User update(Long id, User obj) {
